refactor: ActivityHeader 내의 드롭다운만 클라이언트 컴포넌트로 분리
