+++ src/components/organisms/activity-header/ActivityHeader.tsx
@@ -1,15 +1,7 @@
-'use client';
-
 import AddressLabel from '@/components/atoms/address-label/AddressLabel';
-import Link from 'next/link';
 import Image from 'next/image';
-import { useState } from 'react';
-import { deleteActivity } from '@/queries/activities/delete-activity';
-import { toast } from 'react-toastify';
-import { useRouter } from 'next/navigation';
-import { revalidate } from '@/lib/revalidate';
-import useUser from '@/store/useUser';
-
+import ActivityOptionDropdown from '@/components/molecules/activity-option-dropdown/ActivityOptionDropdown';
+import { getCookie } from '@/app/(action)/(cookie)/cookie';
 interface Props {
   category: string;
   title: string;
@@ -20,7 +12,7 @@ interface Props {
   creatorId: number;
 }
 
-export default function ActivityHeader({
+export default async function ActivityHeader({
   category,
   title,
   rating,
@@ -29,27 +21,14 @@ export default function ActivityHeader({
   activityId,
   creatorId,
 }: Props) {
-  const [optionOpen, setOptionOpen] = useState(false);
-  const handleOptionClick = () => {
-    setOptionOpen((prev) => !prev);
-  };
-  const { user } = useUser();
+  const userId = await getCookie('userId');
 
-  const router = useRouter();
+  let createdByMe = false;
 
-  const createdByMe = creatorId === user?.id;
+  if (userId) {
+    createdByMe = creatorId.toString() === userId;
+  }
 
-  const handleDelete = async () => {
-    const message: any = await deleteActivity(activityId);
-
-    if (message === '삭제가 완료되었습니다') {
-      await revalidate('/');
-      toast(message);
-      router.push('/');
-    } else {
-      toast(message);
-    }
-  };
   return (
     <div className="flex w-full justify-between p-[16px] text-[#112211] md:p-[24px]">
       <div className="flex flex-col">
@@ -65,27 +44,7 @@ export default function ActivityHeader({
           <AddressLabel address={address} />
         </div>
       </div>
-      {createdByMe && (
-        <div className="flex items-center justify-center">
-          <button onClick={handleOptionClick}>
-            <Image src="/meatball-icon.svg" width={40} height={40} alt="options" />
-          </button>
-          <div className="h-1pxr w-1pxr">
-            {optionOpen && (
-              <div className="border-var-gray-6 font-500 relative right-[180px] top-[20px] z-20 w-160pxr rounded-[6px] border bg-white text-18pxr">
-                <Link href={`/myactivity/${activityId}/edit`}>
-                  <button className="border-var-gray-6 w-full shrink-0 border-b px-46pxr py-18pxr">
-                    수정하기
-                  </button>
-                </Link>
-                <button onClick={handleDelete} className="w-full px-46pxr py-18pxr">
-                  삭제하기
-                </button>
-              </div>
-            )}
-          </div>
-        </div>
-      )}
+      {createdByMe && <ActivityOptionDropdown activityId={activityId} />}
     </div>
   );
 }
